+++ src/main/java/com/jelly/mightyminerv2/feature/impl/BlockMiner/BlockMiner.java
@@ -32,7 +32,7 @@ import java.util.Map;
 
 /**
  * BlockMiner
- *
+ * 
  * Main controller class for automatic block mining feature.
  * Implements a state machine pattern to manage different phases of the mining process.
  * Handles mining block selection, breaking, and speed boost management.
@@ -61,7 +61,7 @@ public class BlockMiner extends AbstractFeature {
     @Getter
     @Setter
     private PickaxeAbilityState pickaxeAbilityState;
-
+    
     @Getter
     @Setter
     private BlockMinerError error = BlockMinerError.NONE;
@@ -84,7 +84,7 @@ public class BlockMiner extends AbstractFeature {
 
     @Getter
     private Map<Integer, Integer> blockPriority = new HashMap<>(); // State ID of block -> priority
-
+    
     @Getter
     @Setter
     private BlockPos targetBlockPos; // BlockPos of current block being mined
@@ -92,7 +92,7 @@ public class BlockMiner extends AbstractFeature {
     @Getter
     @Setter
     private Block targetBlockType; // Type of current block being mined
-
+    
     @Getter
     @Setter
     private int miningSpeed;  // Mining speed modifier (affects block breaking time)
@@ -101,8 +101,6 @@ public class BlockMiner extends AbstractFeature {
     @Setter
     private int waitThreshold;  // Stop the macro automatically if it cannot find blocks within the time limit (in ms)
 
-    private boolean particleSpawned = true;
-    private Vec3 particleTarget = null;
 
     @Override
     public String getName() {
@@ -111,7 +109,7 @@ public class BlockMiner extends AbstractFeature {
 
     /**
      * Starts the BlockMiner with specified parameters.
-     *
+     * 
      * @param blocksToMine Array of mineable block types to target
      * @param miningSpeed Base mining speed (higher = faster)
      * @param priority Array of priority values for block selection
@@ -125,7 +123,7 @@ public class BlockMiner extends AbstractFeature {
             this.stop();
             return;
         }
-
+        
         // Validate blocks to mine
         if (blocksToMine == null || Arrays.stream(priority).allMatch(i -> i == 0)) {
             logError("Target blocks not set!");
@@ -139,14 +137,14 @@ public class BlockMiner extends AbstractFeature {
                 blockPriority.put(j, priority[i]);
             }
         }
-
+        
         // Initialize parameters
         this.miningSpeed = miningSpeed - 200;  // Base adjustment to mining speed
         this.enabled = true;
         this.error = BlockMinerError.NONE;
         this.pickaxeAbilityState = PickaxeAbilityState.AVAILABLE;
         this.retryActivatePickaxeAbility = 0;
-
+        
         // Initialize with starting state
         this.currentState = new StartingState();
         this.start();
@@ -173,13 +171,6 @@ public class BlockMiner extends AbstractFeature {
 
         BlockMinerState nextState = currentState.onTick(this);
         transitionTo(nextState);
-
-        if (this.particleSpawned && this.particleTarget != null) {
-            RotationHandler.getInstance().stopFollowingTarget();
-            RotationHandler.getInstance().easeTo(new RotationConfiguration(new Target(this.particleTarget), 400, null).followTarget(true));
-            this.timer.schedule(5000);
-            this.particleSpawned = false;
-        }
 
         if (retryActivatePickaxeAbility >= 4) {
             setError(BlockMinerError.NO_PICKAXE_ABILITY);
@@ -227,34 +218,5 @@ public class BlockMiner extends AbstractFeature {
         }
     }
 
-
-    @SubscribeEvent
-    public void onParticleSpawned(SpawnParticleEvent event) {
-        if (MightyMinerConfig.precisionMiner){
-            if (event.getParticleTypes() == EnumParticleTypes.CRIT && mc.thePlayer.getPositionVector().squareDistanceTo(event.getPos()) < 64) {
-                this.particleSpawned = true;
-                this.particleTarget = event.getPos();
-                RotationHandler.getInstance().queueRotation(
-                        new RotationConfiguration(
-                                new Target(particleTarget),
-                                MightyMinerConfig.getRandomRotationTime() - 3L,
-                                null
-                        )
-                );
-            }
-        }
-    }
-
-    @SubscribeEvent
-    protected void onRender(RenderWorldLastEvent event) {
-        if (targetBlockPos == null) {
-            return; // No target block to render
-        }
-        RenderUtil.drawBlock(targetBlockPos, new Color(0, 255, 0, 100));
-
-        if (this.particleTarget != null) {
-            RenderUtil.drawPoint(this.particleTarget, new Color(255, 0, 0, 100));
-        }
-    }
 
 }
